Let the is_online-function search for the specified username in roomlist
Previously the function would retrieve 500 rooms and check if the specified username is part of it and is online.
This not only is way too much data, this also doesn't guarantee to have the user, we are looking for, in it. Leading to the recording not starting.

I changed the code to use a different API call to only search by username and let chaturbate do the computing.